feat: implement TodoOwnershipGuard to enforce ownership checks for todo operations

+++ src/todo/todo.controller.ts
@@ -11,6 +11,7 @@ import {
   HttpStatus,
   HttpCode,
   ParseIntPipe,
+  ForbiddenException,
 } from "@nestjs/common";
 import { TodoService } from "./todo.service";
 import { CreateTodoDto } from "./dto/createTodo.dto";
@@ -19,6 +20,7 @@ import ApiResponse from "src/utils/ApiResponse";
 import { GetTodosQuery } from "./dto/getTodosQuery.dto";
 import { JwtAuthGuard } from "@/auth/guard/jwt-auth.guard";
 import { CurrentUser } from "@/common/decorators/current-user.decorators";
+import { TodoOwnershipGuard } from "./guards/todo-ownership.guard";
 
 @Controller({
   version: "1",
@@ -40,8 +42,20 @@ export class TodoController {
 
   @Get("/:id")
   @HttpCode(HttpStatus.OK)
-  async getTodoById(@Param("id", ParseIntPipe) id: number) {
+  @UseGuards(TodoOwnershipGuard)
+  async getTodoById(
+    @Param("id", ParseIntPipe) id: number,
+    @CurrentUser() currentUser: any,
+  ) {
     const todo = await this.todoService.getTodoById(id);
+
+    // Check if the logged-in user is the owner of the todo
+    if (todo.user.id !== currentUser.userId) {
+      throw new ForbiddenException(
+        "You are not authorized to access this todo.",
+      );
+    }
+
     return new ApiResponse(HttpStatus.OK, "Todo fetched successfully", todo);
   }
 
@@ -61,10 +75,21 @@ export class TodoController {
 
   @Put("/:id")
   @HttpCode(HttpStatus.OK)
+  @UseGuards(TodoOwnershipGuard)
   async updateTodo(
     @Param("id", ParseIntPipe) id: number,
     @Body() todo: UpdateTodoDto,
+    @CurrentUser() currentUser: any,
   ) {
+    const existingTodo = await this.todoService.getTodoById(id);
+
+    // Check if the logged-in user is the owner of the todo
+    if (existingTodo.user.id !== currentUser.userId) {
+      throw new ForbiddenException(
+        "You are not authorized to update this todo.",
+      );
+    }
+
     const updatedTodo = await this.todoService.updateTodo(id, todo);
     return new ApiResponse(
       HttpStatus.OK,
@@ -75,7 +100,20 @@ export class TodoController {
 
   @Delete("/:id")
   @HttpCode(HttpStatus.OK)
-  async deleteTodo(@Param("id", ParseIntPipe) id: number) {
+  @UseGuards(TodoOwnershipGuard)
+  async deleteTodo(
+    @Param("id", ParseIntPipe) id: number,
+    @CurrentUser() currentUser: any,
+  ) {
+    const existingTodo = await this.todoService.getTodoById(id);
+
+    // Check if the logged-in user is the owner of the todo
+    if (existingTodo.user.id !== currentUser.userId) {
+      throw new ForbiddenException(
+        "You are not authorized to delete this todo.",
+      );
+    }
+
     await this.todoService.deleteTodoById(id);
     return new ApiResponse(HttpStatus.OK, "Todo deleted successfully", {
       todoId: id,
